+++ R/globals.R
@@ -115,7 +115,6 @@ globalVariables(
     "v_repeated",
     "repeated_value",
     "zipcode",
-    "assigned_data",
     "n_shift",
     "num_lines",
     "bags",