+++ R/pullErrors.R
@@ -13,7 +13,7 @@
 #' @param field Field that should be pulled. One of the fields from the following list may be supplied:
 #' \itemize{
 #' \item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_yr, email}
-#' @param distinct If FALSE, returns all error values; if TRUE (default), only returns unique values.
+#' @param unique If FALSE, returns all error values; if TRUE (default), only returns unique values.
 #'
 #' @author Abby Walter, \email{abby_walter@@fws.gov}
 #' @references \url{https://github.com/USFWS/migbirdHIP}
@@ -21,13 +21,13 @@
 #' @export
 
 pullErrors <-
-  function(x, field, distinct = TRUE){
+  function(x, field, unique = TRUE){
 
     # Fail if incorrect field supplied
     stopifnot("Error: Incorrect value supplied for `field` parameter. Please supply one of: title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_yr, email." = field %in% c("title", "firstname", "middle", "lastname", "suffix", "address", "city", "state", "zip", "birth_date", "issue_date", "hunt_mig_birds", "ducks_bag", "geese_bag", "dove_bag", "woodcock_bag", "coots_snipe", "rails_gallinules", "cranes", "band_tailed_pigeon", "brant", "seaducks", "registration_yr", "email"))
 
-    # Fail if incorrect distinct supplied
-    stopifnot("Error: Please supply TRUE or FALSE for `distinct` parameter." = distinct %in% c(TRUE, FALSE, T, F))
+    # Fail if incorrect unique supplied
+    stopifnot("Error: Please supply TRUE or FALSE for `unique` parameter." = unique %in% c(TRUE, FALSE, T, F))
 
     acceptable_fields <-
       names(x)[match("title", names(x)):match("email", names(x))]
@@ -52,7 +52,7 @@ pullErrors <-
 
       } else {
 
-        if(distinct == TRUE) {
+        if(unique == TRUE) {
           pulled_error <- distinct(pulled_error)
         }